# api/openapi.yaml
openapi: 3.0.3

info:
  title: Library API
  version: 0.1.0
  description: |
    Manages books and a history of the changes made to them.

    Single resources are returned inside a `data` envelope. Collections add
    pagination fields alongside `data`. Errors are always `{"error": "..."}`,
    and validation failures add a `fields` map keyed by the offending input.

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Books
    description: Create, read and update books
  - name: System
    description: Operational endpoints

paths:
  /heartbeat:
    get:
      tags: [System]
      summary: Liveness probe
      responses:
        "200":
          description: The service is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
                required: [status, timestamp]

  /books:
    get:
      tags: [Books]
      summary: List books
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A page of books
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginationMeta"
                  - type: object
                    required: [data]
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/BookResponse"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Books]
      summary: Create a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookRequest"
      responses:
        "201":
          description: The created book
          headers:
            Location:
              description: Path of the created book
              schema:
                type: string
                example: /books/0199a1b2-c3d4-7e5f-8a9b-0c1d2e3f4a5b
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/BookResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"

  /books/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Book identifier
        schema:
          type: string
        example: 0199a1b2-c3d4-7e5f-8a9b-0c1d2e3f4a5b

    get:
      tags: [Books]
      summary: Fetch a book
      responses:
        "200":
          description: The book
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/BookResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [Books]
      summary: Replace a book
      description: |
        Replaces every field of the book. Fields omitted from the body are
        treated as cleared, not left unchanged, and each difference from the
        stored book is recorded in its change history.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookRequest"
      responses:
        "200":
          description: The updated book
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: "#/components/schemas/BookResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  parameters:
    Page:
      name: page
      in: query
      description: 1-based page number
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      description: Items per page, clamped to the maximum
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10

  schemas:
    BookRequest:
      type: object
      required: [title, published_on, authors]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          example: The Hobbit
        description:
          type: string
          maxLength: 2000
          example: There and back again.
          x-go-type-skip-optional-pointer: true
        published_on:
          type: string
          format: date
          description: Calendar date, YYYY-MM-DD. Must not be in the future.
          example: "1937-09-21"
          # Decoded as a plain string so an unparseable date is reported as a
          # field error rather than failing the whole body decode.
          x-go-type: string
        authors:
          type: array
          minItems: 1
          description: Author names. Must be non-empty and free of duplicates.
          items:
            type: string
            minLength: 1
            maxLength: 200
          example: [J.R.R. Tolkien]

    BookResponse:
      type: object
      required: [id, title, description, published_on, authors, created_at, updated_at]
      properties:
        id:
          type: string
          example: 0199a1b2-c3d4-7e5f-8a9b-0c1d2e3f4a5b
        title:
          type: string
          example: The Hobbit
        description:
          type: string
          example: There and back again.
        published_on:
          type: string
          format: date
          example: "1937-09-21"
          x-go-type: string
        authors:
          type: array
          items:
            type: string
          example: [J.R.R. Tolkien]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    PaginationMeta:
      type: object
      description: |
        Pagination fields returned alongside `data` on every collection
        response. Composed with an endpoint's item array using allOf.
      required: [total, page, page_size, total_pages]
      properties:
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 10
        total_pages:
          type: integer
          example: 5

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: validation failed
        fields:
          type: object
          description: Present only on validation failures, keyed by input field.
          additionalProperties:
            type: string
          example:
            title: is required
            authors[1]: duplicates "J.R.R. Tolkien"

  responses:
    BadRequest:
      description: |
        The body could not be decoded, or failed validation. Decode failures
        return `error` alone; validation failures also return `fields`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No book with that identifier
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PayloadTooLarge:
      description: The request body exceeded the configured limit
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"