# internal/infrastructure/httpserver/dto/openapi_types.gen.go
// Package dto provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package dto

import (
	"time"
)

// BookRequest defines model for BookRequest.
type BookRequest struct {
	// Authors Author names. Must be non-empty and free of duplicates.
	//
	// Example: ["J.R.R. Tolkien"]
	Authors []string `json:"authors"`

	// Description Example: There and back again.
	Description string `json:"description,omitempty"`

	// PublishedOn Calendar date, YYYY-MM-DD. Must not be in the future.
	//
	// Example: 1937-09-21
	PublishedOn string `json:"published_on"`

	// Title Example: The Hobbit
	Title string `json:"title"`
}

// BookResponse defines model for BookResponse.
type BookResponse struct {
	// Authors Example: ["J.R.R. Tolkien"]
	Authors   []string  `json:"authors"`
	CreatedAt time.Time `json:"created_at"`

	// Description Example: There and back again.
	Description string `json:"description"`

	// ID Example: 0199a1b2-c3d4-7e5f-8a9b-0c1d2e3f4a5b
	ID string `json:"id"`

	// PublishedOn Example: 1937-09-21
	PublishedOn string `json:"published_on"`

	// Title Example: The Hobbit
	Title     string    `json:"title"`
	UpdatedAt time.Time `json:"updated_at"`
}

// Error defines model for Error.
type Error struct {
	// Error Example: validation failed
	Error string `json:"error"`

	// Fields Present only on validation failures, keyed by input field.
	//
	// Example: {"authors[1]":"duplicates \"J.R.R. Tolkien\"","title":"is required"}
	Fields *map[string]string `json:"fields,omitempty"`
}

// PaginationMeta Pagination fields returned alongside `data` on every collection
// response. Composed with an endpoint's item array using allOf.
type PaginationMeta struct {
	// Page Example: 1
	Page int `json:"page"`

	// PageSize Example: 10
	PageSize int `json:"page_size"`

	// Total Example: 42
	Total int `json:"total"`

	// TotalPages Example: 5
	TotalPages int `json:"total_pages"`
}

// Page defines model for Page.
type Page = int

// PageSize defines model for PageSize.
type PageSize = int

// BadRequest defines model for BadRequest.
type BadRequest = Error

// InternalError defines model for InternalError.
type InternalError = Error

// NotFound defines model for NotFound.
type NotFound = Error

// PayloadTooLarge defines model for PayloadTooLarge.
type PayloadTooLarge = Error

// GetBooksParams defines parameters for GetBooks.
type GetBooksParams struct {
	// Page 1-based page number
	Page *Page `form:"page,omitempty" json:"page,omitempty"`

	// PageSize Items per page, clamped to the maximum
	PageSize *PageSize `form:"page_size,omitempty" json:"page_size,omitempty"`
}

// PostBooksJSONRequestBody defines body for PostBooks for application/json ContentType.
type PostBooksJSONRequestBody = BookRequest

// PutBooksIDJSONRequestBody defines body for PutBooksID for application/json ContentType.
type PutBooksIDJSONRequestBody = BookRequest